integration-demos/slack-bot/requirements.txt
bedrock-agentcore
strands-agents
boto3
slack-sdk
